Add inflatable pages and update user experience
Expands the selection of pages for the inflatable section of the site.
Tab functionality is now included in some pages with multiple
"sub-inflatables". Most of the pages utilize minimal JavaScript for
handling tab switching. The exception is the 'surf.php' page, which
is designed to load completely from JavaScript. This feature is
experimental, and may be removed from the final product if the client
is not okay with the decision to move to scripting instead of parsing.

Add a dynamic breadcrumb trail so that when a user navigates to a page
via a specific card from the 'inflatables.php' page it will display the
name of the card they clicked on rather than the page that card's item
is stored on (Harley the Hog shows Harley the Hog despite being stored
on the Mechanical Bull page).